1. 간편한 Log 생성 및 등록을 위한 Func파일 추가, DebugLog 함수 작성
2. PathMgr의 Render에서 로그 호출로 Solution / Project / Content 경로 확인

3. 상위 경로 생성 함수 작성 완료

+++ GameClient/CTimeMgr.cpp
@@ -53,5 +53,5 @@ void CTimeMgr::tick()
 void CTimeMgr::render()
 {
 	CEngine* pEngine = CEngine::GetInst();
-	TextOut(DC, pEngine->GetResolution().x - 250, 10, szBuff, wcslen(szBuff));
+	TextOut(DC, (int)pEngine->GetResolution().x - 250, 10, szBuff, (int)wcslen(szBuff));
 }
